Weight Correction Task

+++ 0x04-python-more_data_structures/100-weight_average.py
@@ -10,6 +10,6 @@ def weight_average(my_list=[]):
             a, b= item
             tup_add += (a * b)
             weight_add += b
-        return tup_add / weight_add
+        return float(tup_add / weight_add)
     else:
         return 0
